requirements/ci.txt
-r base.txt

# documentation, testing and debuging packages
Sphinx==1.2
coverage==3.7.1
django-debug-toolbar
django-discover-runner==1.0
django-extensions
django-nose==1.2
flake8==2.1.0
mock==1.0.1
nose==1.3.0
pep8==1.4.6
pip-tools==0.3.4
